+++ include/ntw_math.h
@@ -30,7 +30,7 @@
  * @param vector The vector \f$\vec{V}\f$, must be at least @a n elements long. 
  * @param c The constant \f$c\f$.
  */
-void NTWM_multDV(const uint64_t n, double vector[static n], const double c);
+void NTWM_multDV(const uint32_t n, double vector[static n], const double c);
 
 /**
  * @brief Does the addition: \f$\vec{A} = \vec{A}+\vec{B}\f$
@@ -39,7 +39,7 @@ void NTWM_multDV(const uint64_t n, double vector[static n], const double c);
  * @param vectorA The vector \f$\vec{A}\f$.
  * @param vectorB The vector \f$\vec{B}\f$.
  */
-void NTWM_addDV(const uint64_t n, double vectorA[static n], const double vectorB[static n]);
+void NTWM_addDV(const uint32_t n, double vectorA[static n], const double vectorB[static n]);
 
 /**
  * @brief Does the substitution: \f$\vec{A} = \vec{A}-\vec{B}\f$
@@ -48,7 +48,7 @@ void NTWM_addDV(const uint64_t n, double vectorA[static n], const double vectorB
  * @param vectorA The vector \f$\vec{A}\f$.
  * @param vectorB The vector \f$\vec{B}\f$.
  */
-void NTWM_subDV(const uint64_t n, double vectorA[static n], const double vectorB[static n]);
+void NTWM_subDV(const uint32_t n, double vectorA[static n], const double vectorB[static n]);
 
 /**
  * @brief Calculates the dot product of @a vectorA and @a vectorB.
@@ -61,7 +61,7 @@ void NTWM_subDV(const uint64_t n, double vectorA[static n], const double vectorB
  * @param vectorB The vector \f$\vec{B}\f$
  * @return double The dot product \f$\vec{A} \cdot \vec{B}\f$
  */
-double NTWM_dotDV(const uint64_t n, const double vectorA[static n], const double vectorB[static n]);
+double NTWM_dotDV(const uint32_t n, const double vectorA[static n], const double vectorB[static n]);
 
 /**
  * @brief Returns the squered magnitude of the @a vector.
@@ -70,7 +70,7 @@ double NTWM_dotDV(const uint64_t n, const double vectorA[static n], const double
  * @param vector The vector \f$\vec{V}\f$
  * @return double \f$\|\vec{V}\|^{2}\f$
  */
-double NTWM_sqMagnDV(const uint64_t n, const double vector[static n]);
+double NTWM_sqMagnDV(const uint32_t n, const double vector[static n]);
 
 /**
  * @brief Normalizes the @a vector.
@@ -78,7 +78,7 @@ double NTWM_sqMagnDV(const uint64_t n, const double vector[static n]);
  * @param n The size of the @a vector.
  * @param vector 
  */
-void NTWM_normalizeDV(const uint64_t n, double vector[static n]);
+void NTWM_normalizeDV(const uint32_t n, double vector[static n]);
 
 /**
  * @brief Returns \f$\sum_{i=0}^{n}V(ind[i])\f$
@@ -88,7 +88,7 @@ void NTWM_normalizeDV(const uint64_t n, double vector[static n]);
  * @param vector The vector \f$\vec{V}\f$
  * @return double The partial sum of the @a vector elements.
  */
-double NTWM_partialSumDV(const uint64_t n, const uint64_t indices[static n], const double vector[static 1]);
+double NTWM_partialSumDV(const uint32_t n, const uint32_t indices[static n], const double vector[static 1]);
 
 /**
  * @brief Assign the values of vectorB to vectorA
@@ -98,7 +98,7 @@ double NTWM_partialSumDV(const uint64_t n, const uint64_t indices[static n], con
  * @param vectorA The \f$\vec{A}\f$.
  * @param vectorB The \f$\vec{B}\f$.
  */
-void NTWM_assignDV(const uint64_t n, double vectorA[static n], const double vectorB[static n]);
+void NTWM_assignDV(const uint32_t n, double vectorA[static n], const double vectorB[static n]);
 
 /**
  * @brief Changes the sequence of the elements of @a vector based on @a sequence..
@@ -120,7 +120,7 @@ void NTWM_reshapeDV(const uint32_t n, double vector[static n], const uint32_t se
  * @param n The size of the vector.
  * @return double* Pointer to the vector: \f$\overbrace{\begin{pmatrix}0 & 0 & \cdots &0\end{pmatrix}}^{n}\f$.
  */
-double* NTWM_newZeroVectorD(const uint64_t n);
+double* NTWM_newZeroVectorD(const uint32_t n);
 
 /**
  * @brief Creates a @a n legth vector with all elements set to @a value.
@@ -131,7 +131,7 @@ double* NTWM_newZeroVectorD(const uint64_t n);
  * @param value The value \f$v\f$ of each element.
  * @return double* Pointer to the vector: \f$\overbrace{\begin{pmatrix}v & v & \cdots &v\end{pmatrix}}^{n}\f$.
  */
-double* NTWM_newUniVectorD(const uint64_t n, const double value);
+double* NTWM_newUniVectorD(const uint32_t n, const double value);
 
 /**
  * @brief Create a copy/subcopy of the @a vector.
@@ -142,7 +142,7 @@ double* NTWM_newUniVectorD(const uint64_t n, const double value);
  * @param vector 
  * @return double* Pointer to the new copy.
  */
-double* NTWM_newCopyDV(const uint64_t n, const double vector[static n]);
+double* NTWM_newCopyDV(const uint32_t n, const double vector[static n]);
 
 /**
  * @brief Prints a double @a vector to the @a stream.
@@ -152,6 +152,6 @@ double* NTWM_newCopyDV(const uint64_t n, const double vector[static n]);
  * @param vector The vector that will be printed.
  * @param decimalDigits The accuracy of the printed elements.
  */
-void NTWM_printDV(FILE* restrict stream, const uint64_t n, const double vector[static n], const unsigned decimalDigits);
+void NTWM_printDV(FILE* restrict stream, const uint32_t n, const double vector[static n], const unsigned decimalDigits);
 
 #endif
